LayoutTests/imported/w3c/web-platform-tests/loading/early-hints/referrer-policy-unsafe-url.h2.window-expected.txt

CONSOLE MESSAGE: <link rel=preload> has an invalid `href` value

FAIL Referrer policy: unsafe-url assert_true: expected true got false
